artifacts/ntbee-flask/requirements.txt
flask>=3.0.0
flask-login>=0.6.3
flask-wtf>=1.2.1
flask-sqlalchemy>=3.1.1
python-dotenv>=1.0.0
werkzeug>=3.0.0
email-validator>=2.1.0
gunicorn>=21.0.0
